requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
fastapi==0.115.11
filelock==3.18.0
h11==0.14.0
idna==3.10
mysql-connector-python==9.2.0
platformdirs==4.3.7
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.29.3
